add popup confirmation windows

+++ www/js/controllers.js
@@ -221,19 +221,31 @@ angular.module('conFusion.controllers', [])
 
 }])
 
-.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+.controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory',
+  'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout',
+  function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
     $scope.baseURL = baseURL;
     $scope.shouldShowDelete = false;
+
+    $ionicLoading.show({
+        template: '<ion-spinner></ion-spinner> Loading...'
+    });
 
     $scope.favorites = favoriteFactory.getFavorites();
 
     $scope.dishes = menuFactory.getDishes().query(
         function (response) {
             $scope.dishes = response;
+            $timeout(function () {
+                $ionicLoading.hide();
+            }, 1000);
         },
         function (response) {
             $scope.message = "Error: " + response.status + " " + response.statusText;
+            $timeout(function () {
+                $ionicLoading.hide();
+            }, 1000);
         });
     console.log($scope.dishes, $scope.favorites);
 
@@ -244,7 +256,21 @@ angular.module('conFusion.controllers', [])
 
     $scope.deleteFavorite = function (index) {
 
-        favoriteFactory.deleteFromFavorites(index);
+      var confirmPopup = $ionicPopup.confirm({
+          title: 'Confirm Delete',
+          template: 'Are you sure you want to delete this item?'
+      });
+
+      confirmPopup.then(function (res) {
+          if (res) {
+              console.log('Ok to delete');
+              favoriteFactory.deleteFromFavorites(index);
+          } else {
+              console.log('Canceled delete');
+          }
+      });
+
+      $scope.shouldShowDelete = false;
 
     }
 }])
